requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
Pillow>=10.0.0
grad-cam>=1.4.8
kaggle>=1.5.16
scipy>=1.11.0
